map_model::city
Struct City

pub struct City {
    pub name: CityName,
    pub boundary: Polygon,
    pub areas: Vec<(AreaType, Polygon)>,
    pub districts: Vec<(MapName, Polygon)>,
}

A single city (like Seattle) can be broken down into multiple boundary polygons (udistrict, ballard, downtown, etc). The load map screen uses this struct to display the entire city.

Fields
name: CityNameboundary: Polygonareas: Vec<(AreaType, Polygon)>districts: Vec<(MapName, Polygon)>

The individual maps

Implementations

impl City

pub fn from_huge_map(huge_map: &Map) -> City

If there’s a single map covering all the smaller maps, use this.

pub fn from_individual_maps(city_name: &CityName, timer: &mut Timer<'_>) -> City

Generate a city from a bunch of smaller, individual maps. The boundaries of those maps may overlap and may have gaps between them.