Module map_model::connectivity

Modules

walking

Structs

WalkingOptions

Enums

WalkingNode

Functions

all_vehicle_costs_from
Starting from one building, calculate the cost to all others. If a destination isn't reachable, it won't be included in the results. Ignore results greater than the time_limit away.

all_walking_costs_from
Starting from one building, calculate the cost to all others. If a destination isn't reachable, it won't be included in the results. Ignore results greater than the time_limit away.

driving_cost
Different unit based on constraints.

find_scc
Calculate the srongy connected components (SCC) of the part of the map accessible by constraints (ie, the graph of sidewalks or driving+bike lanes). The largest component is the "main" graph; the rest is disconnected. Returns (lanes in the largest "main" component, all other disconnected lanes)