Function map_model::connectivity::all_walking_costs_from

pub fn all_walking_costs_from(
    map: &Map,
    starts: Vec<Spot>,
    time_limit: Duration,
    opts: WalkingOptions
) -> HashMap<BuildingID, Duration>

Starting from some initial buildings, calculate the cost to all others. If a destination isn’t reachable, it won’t be included in the results. Ignore results greater than the time_limit away.

If all of the start buildings are on the shoulder of a road and !opts.allow_shoulders, then the results will always be empty.