map_model::connectivity
Function vehicle_cost

pub fn vehicle_cost(
    dr: DirectedRoadID,
    mvmnt: MovementID,
    constraints: PathConstraints,
    params: &RoutingParams,
    map: &Map,
) -> Option<Duration>

This returns the pathfinding cost of crossing one road and turn, in units of time. It factors in the ideal time to cross the space and penalties for entering an access-restricted zone, taking an unprotected turn, or going up a steep hill for some vehicle types. If this returns None, then the movement isn’t actually allowed.