map_model

Module edits

Once a Map exists, the player can edit it in the UI (producing MapEdits in-memory), then save the changes to a file (as PermanentMapEdits). See https://a-b-street.github.io/docs/tech/map/edits.html.

Re-exports

pub use self::perma::PermanentMapEdits;

Modules

apply 🔒

compat 🔒

perma 🔒

perma_traffic_signal

A representation of traffic signal configuration that references OpenStreetMap IDs and is hopefully robust to minor edits over time.

Structs

EditEffects

EditIntersection

EditRoad

MapEdits

Represents changes to a map. Note this isn’t serializable – that’s what PermanentMapEdits does.

Enums

EditCmd

EditIntersectionControl