map_model::edits
Module apply

Functions

fix_building_driveways 🔒

Recalculate the driveways of some buildings after map edits.

fix_parking_lot_driveways 🔒

Recalculate the driveways of some parking lots after map edits.

modify_lanes 🔒

recalculate_intersection_polygon 🔒

recalculate_turns 🔒