Enum map_model::CommonEndpoint

pub enum CommonEndpoint {
    One(IntersectionID),
    Both,
    None,
}

Variants

One(IntersectionID)

Two lanes/roads share one endpoint

Both

Two lanes/roads share both endpoints, because they both belong to the same road, or there are two different roads connecting the same pair of intersections

None

Two lanes/roads don’t have any common endpoints

Implementations

impl CommonEndpoint

pub fn new( obj1: (IntersectionID, IntersectionID), obj2: (IntersectionID, IntersectionID) ) -> CommonEndpoint