map_model

Enum PathConstraints

pub enum PathConstraints {
    Pedestrian,
    Car,
    Bike,
    Bus,
    Train,
}

Who’s asking for a path?

Variants

Pedestrian

Car

Bike

Bus

Train

Implementations

impl PathConstraints

pub fn all() -> Vec<PathConstraints>

pub fn from_lt(lt: LaneType) -> PathConstraints

Not bijective, but this is the best guess of user intent

pub fn can_use(self, lane: &Lane, map: &Map) -> bool

Can an agent use a lane? There are some subtle exceptions with using bus-only lanes for turns.

pub fn can_use_road(self, road: &Road, map: &Map) -> bool

Can an agent use a road in either direction? There are some subtle exceptions with using bus-only lanes for turns.

pub(crate) fn filter_lanes(self, choices: Vec<LaneID>, map: &Map) -> Vec<LaneID>

Strict for bikes. If there are bike lanes, not allowed to use other lanes.