Function map_model::make::bridges::find_bridges

pub fn find_bridges(roads: &mut Vec<Road>, bounds: &Bounds, timer: &mut Timer<'_>)

Look for roads underneath bridges, then lower their z-order. OSM tags bridges and tunnels, but not the roads that pass under bridges.