map_model::make

Function match_points_to_lanes

pub fn match_points_to_lanes<F: Fn(&Lane) -> bool>(
    map: &Map,
    pts: HashSet<HashablePt2D>,
    filter: F,
    buffer: Distance,
    max_dist_away: Distance,
    timer: &mut Timer<'_>,
) -> HashMap<HashablePt2D, Position>

Snap points to an exact Position along the nearest lane. If the result doesn’t contain a requested point, then there was no matching lane close enough.