Function map_model::make::initial::geometry::intersection_polygon

pub fn intersection_polygon(
    intersection_id: NodeID, 
    intersection_roads: BTreeSet<OriginalRoad>, 
    roads: &mut BTreeMap<OriginalRoad, Road>, 
    merged: bool
) -> Result<(Polygon, Vec<(String, Polygon)>)>

Also returns a list of labeled polygons for debugging.

Ideally, the resulting polygon should exist entirely within the thick bands around all original roads – it just carves up part of that space, doesn’t reach past it. But that’s not always true yet.