Function map_model::make::traffic_signals::get_possible_policies

pub fn get_possible_policies(
    map: &Map, 
    id: IntersectionID, 
    enforce_manual_signals: bool
) -> Vec<(String, ControlTrafficSignal)>

Applies a bunch of heuristics to a single intersection, returning the valid results in best-first order. The signal configuration is only based on the roads connected to the intersection.

If enforce_manual_signals is true, then any data from the traffic_signal_data crate that matches the map will be validated against the current map. If the config is out-of-date, this method will panic, so that whoever is running the importer can immediately fix the config.