Module map_model::make::walking_turns

Functions

filter_turns
Filter out crosswalks on really short roads. In reality, these roads are usually located within an intersection, which isn’t a valid place for a pedestrian crossing.

get_sidewalk

make_crosswalks

make_degenerate_crosswalks

make_footway_turns
At an intersection of footpaths only, just generate a turn between every pair of lanes.

make_shared_sidewalk_corner

make_walking_turns
Generate Crosswalk and SharedSidewalkCorner (places where two sidewalks directly meet) turns

make_walking_turns_v2
A complete rewrite of make_walking_turns, which looks at all sidewalks (or lack thereof) in counter-clockwise order around an intersection. Based on adjacency, create a SharedSidewalkCorner or a Crosswalk.

turn_id