Module map_model::objects::block

Structs

Block

A block is defined by a perimeter that traces along the sides of roads. Inside the perimeter, the block may contain buildings and interior roads. In the simple case, a block represents a single “city block”, with no interior roads. It may also cover a “neighborhood”, where the perimeter contains some “major” and the interior consists only of “minor” roads.

Perimeter

A sequence of roads in order, beginning and ending at the same place. No “crossings” – tracing along this sequence should geometrically yield a simple polygon.