Struct map_model::objects::block::Perimeter

pub struct Perimeter {
    pub roads: Vec<RoadSideID>,
    pub interior: BTreeSet<RoadID>,
}

A sequence of roads in order, beginning and ending at the same place. No “crossings” – tracing along this sequence should geometrically yield a simple polygon.

Fields
roads: Vec<RoadSideID>interior: BTreeSet<RoadID>

These roads exist entirely within the perimeter

Implementations

impl Perimeter

pub fn single_block(map: &Map, start: LaneID) -> Result<Perimeter>

Starting at any lane, snap to the nearest side of that road, then begin tracing a single block, with no interior roads. This will fail if a map boundary is reached. The results are unusual when crossing the entrance to a tunnel or bridge.

pub fn find_all_single_blocks(map: &Map) -> Vec<Perimeter>

This calculates all single block perimeters for the entire map. The resulting list does not cover roads near the map boundary.

fn undo_invariant(&mut self)

A perimeter has the first and last road matching up, but that’s confusing to work with. Temporarily undo that.

fn restore_invariant(&mut self)

Restore the first=last invariant. Methods may temporarily break this, but must restore it before returning.

fn try_to_merge(
&mut self,
map: &Map,
other: &mut Perimeter,
debug_failures: bool
) -> bool

Try to merge two blocks. Returns true if this is successful, which will only be when the blocks are adjacent, but the merge wouldn’t create an interior “hole”.

Note this may modify both perimeters and still return false. The modification is just to rotate the order of the road loop; this doesn’t logically change the perimeter.

TODO Due to https://github.com/a-b-street/abstreet/issues/841, it seems like rotation sometimes breaks to_block, so for now, always revert to the original upon failure.

pub fn merge_all(
map: &Map,
input: Vec<Perimeter>,
stepwise_debug: bool
) -> Vec<Perimeter>

Try to merge all given perimeters. If successful, only one perimeter will be returned. Perimeters are never “destroyed” – if not merged, they’ll appear in the results. If stepwise_debug is true, returns after performing just one merge.

pub fn collapse_deadends(&mut self)

If the perimeter follows any dead-end roads, “collapse” them and instead make the perimeter contain the dead-end.

pub fn partition_by_predicate<F: Fn(RoadID) -> bool>(
input: Vec<Perimeter>,
predicate: F
) -> Vec<Vec<Perimeter>>

Consider the perimeters as a graph, with adjacency determined by sharing any road in common. Partition adjacent perimeters, subject to the predicate. Each partition should produce a single result with merge_all.

pub fn calculate_coloring(
input: &[Perimeter],
num_colors: usize
) -> Option<Vec<usize>>

Assign each perimeter one of num_colors, such that no two adjacent perimeters share the same color. May fail. The resulting colors are expressed as [0, num_colors).

pub fn to_block(self, map: &Map) -> Result<Block>

pub fn contains(&self, other: &Perimeter) -> bool

Does this perimeter completely enclose the other?