Module map_model::objects::road

Structs

Crossing

DirectedRoadID

OriginalRoad

Refers to a road segment between two nodes, using OSM IDs. Note OSM IDs are not stable over time and the relationship between a road/intersection and way/node isn’t 1:1 at all.

Road

A Road represents a segment between exactly two Intersections. It contains Lanes as children.

RoadID

RoadSideID

Enums

SideOfRoad

See https://wiki.openstreetmap.org/wiki/Forward_%26_backward,left%26_right.