Module map_model::objects::traffic_signals

Structs

ControlTrafficSignal
A traffic signal consists of a sequence of Stages that repeat in a cycle. Most Stages last for a fixed duration. During a single Stage, some movements are protected (can proceed with the highest priority), while others are permitted (have to yield before proceeding).

Stage

Enums

StageType

Constants

CROSSWALK_PACE

Functions

export_movement

find_r

import_movement