Struct map_model::objects::traffic_signals::ControlTrafficSignal

pub struct ControlTrafficSignal {
    pub id: IntersectionID,
    pub stages: Vec<Stage>,
    pub offset: Duration,
}

A traffic signal consists of a sequence of Stages that repeat in a cycle. Most Stages last for a fixed duration. During a single Stage, some movements are protected (can proceed with the highest priority), while others are permitted (have to yield before proceeding).

Fields
id: IntersectionIDstages: Vec<Stage>offset: Duration
Implementations

impl ControlTrafficSignal

pub fn new(map: &Map, id: IntersectionID) -> ControlTrafficSignal

pub(crate) fn validating_new(
map: &Map,
id: IntersectionID
) -> ControlTrafficSignal

Only call this variant while importing the map, to enforce that baked-in signal config is valid.

pub fn get_possible_policies(
map: &Map,
id: IntersectionID
) -> Vec<(String, ControlTrafficSignal)>

pub fn get_min_crossing_time(&self, idx: usize, i: &Intersection) -> Duration

pub fn validate(&self, i: &Intersection) -> Result<()>

pub fn convert_to_ped_scramble(&mut self, i: &Intersection) -> bool

Move crosswalks from stages, adding them to an all-walk as last stage. This may promote yields to protected. True is returned if any stages were added or modified.

pub fn convert_to_ped_scramble_without_promotion(
&mut self,
i: &Intersection
) -> bool

Move crosswalks from stages, adding them to an all-walk as last stage. This does not promote yields to protected. True is returned if any stages were added or modified.

fn internal_convert_to_ped_scramble(
&mut self,
promote_yield_to_protected: bool,
i: &Intersection
) -> bool

pub fn adjust_major_minor_timing(
&mut self,
major: Duration,
minor: Duration,
map: &Map
) -> Result<()>

Modifies the fixed timing of all stages, applying either a major or minor duration, depending on the relative rank of the roads involved in the intersection. If this transformation couldn’t be applied, returns an error. Even if an error is returned, the signal may have been changed – so only call this on a cloned signal.

pub fn missing_turns(&self, i: &Intersection) -> BTreeSet<MovementID>

pub fn simple_cycle_duration(&self) -> Duration

How long a full cycle of the signal lasts, assuming no actuated timings.

impl ControlTrafficSignal

pub fn export(&self, map: &Map) -> TrafficSignal

pub(crate) fn import(
raw: TrafficSignal,
id: IntersectionID,
map: &Map
) -> Result<ControlTrafficSignal>