Struct map_model::pathfind::v1::PathRequest

pub struct PathRequest {
    pub start: Position,
    pub end: Position,
    pub constraints: PathConstraints,
    pub(crate) alt_start: Option<(Position, Duration)>,
}

Fields
start: Positionend: Positionconstraints: PathConstraintsalt_start: Option<(Position, Duration)>
Implementations

impl PathRequest

pub fn between_buildings(
map: &Map,
from: BuildingID,
to: BuildingID,
constraints: PathConstraints
) -> Option<PathRequest>

Determines the start and end position to travel between two buildings for a certain mode. The path won’t cover modality transfers – if somebody has to walk between the building and a parking spot or bikeable position, that won’t be captured here.

pub fn walking(start: Position, end: Position) -> PathRequest

The caller must pass in two valid sidewalk positions. This isn’t verified.

pub fn vehicle(
start: Position,
end: Position,
constraints: PathConstraints
) -> PathRequest

The caller must pass in two valid positions for the vehicle type. This isn’t verified. No off-side turns from driveways happen; the exact start position is used.

pub fn leave_from_driveway(
start: Position,
end: Position,
constraints: PathConstraints,
map: &Map
) -> PathRequest

The caller must pass in two valid positions for the vehicle type. This isn’t verified. TODO The vehicle may cut exit the driveway onto the off-side of the road.