Module map_model::pathfind::v2

Structures related to the new road-based pathfinding (https://github.com/a-b-street/abstreet/issues/555) live here. When the transition is done, things here will probably move into pathfind/mod.rs.

Structs

PathV2
A path between two endpoints for a particular mode. This representation is immutable and doesn’t prescribe specific lanes and turns to follow.

Enums

PathStepV2
One step along a path.

Functions

find_uber_turns