Module map_model::pathfind::walking

Pathfinding for pedestrians using contraction hierarchies, as well as figuring out if somebody should use public transit.

Structs

SidewalkPathfinder

Enums

WalkingNode

Functions

make_input_graph

one_step_walking_path

transit_input_graph

walking_cost
The cost is time in seconds, rounded to a usize

walking_path_to_steps