Struct map_model::ControlTrafficSignal

pub struct ControlTrafficSignal {
    pub id: IntersectionID,
    pub stages: Vec<Stage>,
    pub offset: Duration,
    pub movements: BTreeMap<MovementID, Movement>,
}

A traffic signal consists of a sequence of Stages that repeat in a cycle. Most Stages last for a fixed duration. During a single Stage, some movements are protected (can proceed with the highest priority), while others are permitted (have to yield before proceeding).

Fields
id: IntersectionIDstages: Vec<Stage>offset: Durationmovements: BTreeMap<MovementID, Movement>
Implementations

impl ControlTrafficSignal

pub fn new(
map: &Map,
id: IntersectionID,
timer: &mut Timer<'_>
) -> ControlTrafficSignal

pub fn get_possible_policies(
map: &Map,
id: IntersectionID,
timer: &mut Timer<'_>
) -> Vec<(String, ControlTrafficSignal)>

pub fn brute_force(map: &Map, id: IntersectionID)

pub fn get_min_crossing_time(&self, idx: usize) -> Duration

pub(crate) fn validate(&self) -> Result<(), String>

pub fn convert_to_ped_scramble(&mut self) -> bool

Returns true if this did anything

pub fn turn_to_movement(&self, turn: TurnID) -> MovementID

pub fn missing_turns(&self) -> BTreeSet<MovementID>

pub fn compressed_id(&self, turn: TurnID) -> CompressedMovementID

pub fn simple_cycle_duration(&self) -> Duration

How long a full cycle of the signal lasts, assuming no actuated timings.

impl ControlTrafficSignal

pub fn export(&self, map: &Map) -> TrafficSignal

pub(crate) fn import(
raw: TrafficSignal,
id: IntersectionID,
map: &Map
) -> Result<ControlTrafficSignal, String>