Struct map_model::MapEdits

pub struct MapEdits {
    pub edits_name: String,
    pub commands: Vec<EditCmd>,
    pub original_roads: BTreeMap<RoadID, EditRoad>,
    pub original_intersections: BTreeMap<IntersectionID, EditIntersection>,
    pub changed_routes: BTreeSet<TransitRouteID>,
    pub proposal_description: Vec<String>,
    pub proposal_link: Option<String>,
}

Represents changes to a map. Note this isn’t serializable – that’s what PermanentMapEdits does.

Fields
edits_name: Stringcommands: Vec<EditCmd>

A stack, oldest edit is first. The same object may be edited multiple times in this stack, until compress() happens.

original_roads: BTreeMap<RoadID, EditRoad>

Derived from commands, kept up to date by update_derived

original_intersections: BTreeMap<IntersectionID, EditIntersection>changed_routes: BTreeSet<TransitRouteID>proposal_description: Vec<String>

Some edits are included in the game by default, in data/system/proposals, as “community proposals.” They require a description and may have a link to a write-up.

proposal_link: Option<String>
Implementations

impl MapEdits

pub fn to_permanent(&self, map: &Map) -> PermanentMapEdits

Encode the edits in a permanent format, referring to more-stable OSM IDs.

impl MapEdits

pub(crate) fn new() -> MapEdits

pub fn load_from_file( map: &Map, path: String, timer: &mut Timer<'_> ) -> Result<MapEdits>

Load map edits from a JSON file. Strip out any commands that’re broken because they don’t match the current map. If the resulting edits are totally empty, consider that a failure – the edits likely don’t cover this map at all.

pub fn load_from_bytes(map: &Map, bytes: Vec<u8>) -> Result<MapEdits>

Load map edits from the given JSON bytes. Strip out any commands that’re broken because they don’t match the current map. If the resulting edits are totally empty, consider that a failure – the edits likely don’t cover this map at all.

fn save(&self, map: &Map)

fn update_derived(&mut self, map: &Map)

pub fn compress(&mut self, map: &Map)

Assumes update_derived has been called.

pub fn changed_lanes(&self, map: &Map) -> (BTreeSet<LaneID>, BTreeSet<RoadID>)

Pick apart changed_roads and figure out if an entire road was edited, or just a few lanes. Doesn’t return deleted lanes.

pub fn get_checksum(&self, map: &Map) -> String

Produces an md5sum of the contents of the edits.

pub fn get_title(&self) -> &str

Get the human-friendly of these edits. If they have a description, the first line is the title. Otherwise we use the filename.

pub fn is_crossing_modified(&self, r: RoadID, crossing: &Crossing) -> bool

pub fn is_filter_modified(&self, r: RoadID, filter: &RoadFilter) -> bool

pub fn is_diagonal_filter_modified( &self, i: IntersectionID, filter: &DiagonalFilter ) -> bool