Struct map_model::PathRequest

pub struct PathRequest {
    pub start: Position,
    pub end: Position,
    pub constraints: PathConstraints,
    pub(crate) alt_start: Option<(Position, Duration)>,
}

Fields
start: Positionend: Positionconstraints: PathConstraintsalt_start: Option<(Position, Duration)>
Implementations

impl PathRequest

pub fn between_buildings( map: &Map, from: BuildingID, to: BuildingID, constraints: PathConstraints ) -> Option<PathRequest>

Determines the start and end position to travel between two buildings for a certain mode. The path won’t cover modality transfers – if somebody has to walk between the building and a parking spot or bikeable position, that won’t be captured here.

pub fn walking(start: Position, end: Position) -> PathRequest

The caller must pass in two valid sidewalk positions. This isn’t verified.

pub fn vehicle( start: Position, end: Position, constraints: PathConstraints ) -> PathRequest

The caller must pass in two valid positions for the vehicle type. This isn’t verified. No off-side turns from driveways happen; the exact start position is used.

pub fn leave_from_driveway( start: Position, end: Position, constraints: PathConstraints, map: &Map ) -> PathRequest

The caller must pass in two valid positions for the vehicle type. This isn’t verified. TODO The vehicle may cut exit the driveway onto the off-side of the road.

pub fn between_directed_roads( map: &Map, from: DirectedRoadID, to: DirectedRoadID, constraints: PathConstraints ) -> Option<PathRequest>

Create a request from the beginning of one road to the end of another. Picks an arbitrary start and end lane from the available ones.

pub fn deduplicate( map: &Map, requests: Vec<PathRequest> ) -> Vec<(PathRequest, usize)>

Group similar requests together, returning the number of matches. This can be used to calculate less paths and multiply whatever’s being measured by the count.

Note this throws away detail. It only groups by the mode and from/to parent. Exact position and alternate starting points are lost.