map_model

Module traversable

Structs

Position

Represents a specific point some distance along a lane.

Enums

Traversable

Either a lane or a turn, where most movement happens.

Constants

MAX_BIKE_SPEED

MAX_WALKING_SPEED

Functions

bike_speed_on_incline 🔒

walking_speed_on_incline 🔒