Function map_model::connectivity::vehicle_cost

pub fn vehicle_cost(
    lane: &Lane, 
    turn: &Turn, 
    constraints: PathConstraints, 
    params: &RoutingParams, 
    map: &Map
) -> Duration

This returns the pathfinding cost of crossing one lane and turn. This is also expressed in units of time. It factors in the ideal time to cross the space, along with penalties for entering an access-restricted zone, taking an unprotected turn, and so on.