Module map_model::edits::perma_traffic_signal

A representation of traffic signal configuration that references OpenStreetMap IDs and is hopefully robust to minor edits over time.

Structs

DirectedRoad

A road segment connecting two intersections, and a direction along the segment.

Plan

A plan describes how a traffic signal is configured during some period of time. Multiple plans allow a single intersection to behave differently in the middle of the night with low traffic, compared to the middle of rush hour.

Stage

A traffic signal is in one stage at any time. The stage describes what movements are possible.

TrafficSignal

Turn

A movement through an intersection.

Enums

StageType

How long a stage lasts before moving to the next one.