map_model

Module make

See https://a-b-street.github.io/docs/tech/map/importing/index.html for an overview. This module covers the RawMap->Map stage.

Re-exports

pub use self::parking_lots::snap_driveway;

Modules

bridges 🔒

buildings 🔒

parking_lots 🔒

traffic_signals

The various traffic signal generators live in the traffic signal module. Eventually, we might want to move to a trait. For now, there’s a single make_traffic_signal static method in each generator file, which is called to generate a traffic signal of a particular flavor.

transit

turns

walking_turns 🔒

Structs

RawToMapOptions

Options for converting RawMaps to Maps.

Functions

match_points_to_lanes

Snap points to an exact Position along the nearest lane. If the result doesn’t contain a requested point, then there was no matching lane close enough.

snap_nodes_to_line 🔒

snap_nodes_with_data_to_line 🔒

trim_path

Adjust the path to start on the polygon’s border, not center.