Module map_model::make

See https://dabreegster.github.io/abstreet/map/importing/index.html for an overview. This module covers the RawMap->Map stage.

Modules

bridges

buildings

initial
Naming is confusing, but RawMap -> InitialMap -> Map. InitialMap is separate pretty much just for the step of producing https://dabreegster.github.io/abstreet/map/importing/geometry.html.

medians

parking_lots

remove_disconnected

traffic_signals

transit

turns

walking_turns

Functions

match_points_to_lanes
Snap points to an exact Position along the nearest lane. If the result doesn't contain a requested point, then there was no matching lane close enough.