Module map_model::make::initial::geometry

OSM describes roads as center-lines that intersect. Turn these into road and intersection polygons roughly by

treating the road as a PolyLine with a width, so that it has a left and right edge

finding the places where the edges of different roads intersect

"Trimming back" the center lines to avoid the overlap

Producing a polygon for the intersection itsef

Structs

Piece

Constants

DEGENERATE_INTERSECTION_HALF_LENGTH

Functions

close_off_polygon

deadend

generalized_trim_back

intersection_polygon
Also returns a list of labeled polygons for debugging.

on_off_ramp