Module map_model::make::turns

Functions

curvey_turn

ensure_unique

expected_turn_types_for_four_way

from_pt

make_all_turns
Generate all driving and walking turns at an intersection, accounting for OSM turn restrictions.

make_vehicle_turns

remove_merging_turns

to_pt

turn_type_from_angles

verify_vehicle_connectivity
Ideally, we want every incoming lane to lead to at least one lane of the same type, and every outgoing lane to be reachable by at least one lane of the same type. But if it’s a bus or bike lane, settle for being connected to anything – even just a driving lane. There’s naturally places where these dedicated lanes start and end.