Struct map_model::objects::intersection::Intersection

pub struct Intersection {
    pub id: IntersectionID,
    pub polygon: Polygon,
    pub turns: BTreeSet<TurnID>,
    pub elevation: Distance,
    pub intersection_type: IntersectionType,
    pub orig_id: NodeID,
    pub incoming_lanes: Vec<LaneID>,
    pub outgoing_lanes: Vec<LaneID>,
    pub roads: BTreeSet<RoadID>,
}

An intersection connects roads. Most have >2 roads and are controlled by stop signs or traffic signals. Roads that lead to the boundary of the map end at border intersections, with only that one road attached.

Fields
id: IntersectionIDpolygon: Polygon

This needs to be in clockwise orientation, or later rendering of sidewalk corners breaks.

turns: BTreeSet<TurnID>elevation: Distanceintersection_type: IntersectionTypeorig_id: NodeIDincoming_lanes: Vec<LaneID>

Note that a lane may belong to both incoming_lanes and outgoing_lanes.

outgoing_lanes: Vec<LaneID>roads: BTreeSet<RoadID>
Implementations

impl Intersection

pub fn is_border(&self) -> bool

pub fn is_incoming_border(&self) -> bool

pub fn is_outgoing_border(&self) -> bool

pub fn is_closed(&self) -> bool

pub fn is_stop_sign(&self) -> bool

pub fn is_traffic_signal(&self) -> bool

pub fn is_light_rail(&self, map: &Map) -> bool

pub fn is_private(&self, map: &Map) -> bool

pub fn is_footway(&self, map: &Map) -> bool

pub fn get_incoming_lanes(
&self,
map: &Map,
constraints: PathConstraints
) -> Vec<LaneID>

pub fn get_outgoing_lanes(
&self,
map: &Map,
constraints: PathConstraints
) -> Vec<LaneID>

Strict for bikes. If there are bike lanes, not allowed to use other lanes.

pub fn get_zorder(&self, map: &Map) -> isize

pub fn get_rank(&self, map: &Map) -> RoadRank

pub fn get_roads_sorted_by_incoming_angle(
&self,
all_roads: &Vec<Road>
) -> Vec<RoadID>

pub fn some_outgoing_road(&self, map: &Map) -> Option<DirectedRoadID>

pub fn some_incoming_road(&self, map: &Map) -> Option<DirectedRoadID>

pub fn name(&self, lang: Option<&String>, map: &Map) -> String