map_model::objects
Module lane

Structs

Lane

A road segment is broken down into individual lanes, which have a LaneType.

LaneID

A lane is identified by its parent road and its position, ordered from the left.

Enums

CommonEndpoint

Constants

PARKING_LOT_SPOT_LENGTH

From some manually audited cases in Seattle, the length of parallel street parking spots is a bit different than the length in parking lots, so set a different value here.

Functions

parse_turn_type_from_osm ๐