Module map_model::objects::lane

Structs

Lane
A road segment is broken down into individual lanes, which have a LaneType.

LaneID

LaneSpec

Enums

LaneType

Constants

NORMAL_LANE_THICKNESS

PARKING_LOT_SPOT_LENGTH
From some manually audited cases in Seattle, the length of parallel street parking spots is a bit different than the length in parking lots, so set a different value here.

SERVICE_ROAD_LANE_THICKNESS

SHOULDER_THICKNESS

SIDEWALK_THICKNESS