Struct map_model::pathfind::pathfinder::Pathfinder

pub struct Pathfinder {
    car_graph: VehiclePathfinder,
    bike_graph: VehiclePathfinder,
    bus_graph: VehiclePathfinder,
    train_graph: VehiclePathfinder,
    walking_graph: SidewalkPathfinder,
    walking_with_transit_graph: SidewalkPathfinder,
    params: RoutingParams,
}

Fields
car_graph: VehiclePathfinderbike_graph: VehiclePathfinderbus_graph: VehiclePathfindertrain_graph: VehiclePathfinderwalking_graph: SidewalkPathfinderwalking_with_transit_graph: SidewalkPathfinderparams: RoutingParams
Implementations

impl Pathfinder

pub fn empty() -> Pathfinder

Quickly create an invalid pathfinder, just to make borrow checking / initialization order work.

pub fn new(
map: &Map,
params: RoutingParams,
engine: CreateEngine<'_>,
timer: &mut Timer<'_>
) -> Pathfinder

pub fn pathfind(&self, req: PathRequest, map: &Map) -> Option<PathV2>

Finds a path from a start to an end for a certain type of agent.

pub fn pathfind_with_params(
&self,
req: PathRequest,
params: &RoutingParams,
map: &Map
) -> Option<PathV2>

Finds a path from a start to an end for a certain type of agent. May use custom routing parameters.

pub fn all_costs_from(
&self,
req: PathRequest,
map: &Map
) -> Option<(Duration, HashMap<DirectedRoadID, Duration>)>

pub fn should_use_transit(
&self,
map: &Map,
start: Position,
end: Position
) -> Option<(BusStopID, Option<BusStopID>, BusRouteID)>

pub fn apply_edits(&mut self, map: &Map, timer: &mut Timer<'_>)