Module map_model::pathfind::uber_turns

To deal with complicated intersections and short roads in OSM, cluster intersections close together and then calculate UberTurns that string together several turns.

Structs

IntersectionCluster

This only applies to VehiclePathfinder; walking through these intersections is nothing special. And in fact, even lanes only for buses/bikes are ignored.

UberTurn

UberTurnV2

A sequence of movements through a cluster of intersections. Like UberTurn, but at the granularity of directed roads, not lanes.

Functions

flood 🔒

trace_back 🔒