Module map_model::raw

The convert_osm crate produces a RawMap from OSM and other data. Storing this intermediate structure is useful to iterate quickly on parts of the map importing pipeline without having to constantly read .osm files, and to visualize the intermediate state with map_editor.

Structs

OriginalRoad
A way to refer to roads across many maps and over time. Also trivial to relate with OSM to find upstream problems.

RawArea

RawBuilding

RawBusRoute

RawBusStop

RawIntersection

RawMap

RawParkingLot

RawRoad

TurnRestriction

Enums

RestrictionType