Struct map_model::MapConfig

pub struct MapConfig {
    pub driving_side: DrivingSide,
    pub bikes_can_use_bus_lanes: bool,
    pub inferred_sidewalks: bool,
    pub street_parking_spot_length: Distance,
    pub turn_on_red: bool,
    pub osm2lanes: bool,
    pub find_dog_legs_experiment: bool,
    pub merge_osm_ways: Vec<OriginalRoad, Global>,
}

Fields
driving_side: DrivingSide

If true, driving happens on the right side of the road (USA). If false, on the left (Australia).

bikes_can_use_bus_lanes: boolinferred_sidewalks: bool

If true, roads without explicitly tagged sidewalks may be assigned sidewalks or shoulders. If false, no inference will occur and separate sidewalks and crossings will be included.

street_parking_spot_length: Distance

Street parking is divided into spots of this length. 8 meters is a reasonable default, but people in some regions might be more accustomed to squeezing into smaller spaces. This value can be smaller than the hardcoded maximum car length; cars may render on top of each other, but otherwise the simulation doesn’t care.

turn_on_red: bool

If true, turns on red which do not conflict crossing traffic (‘right on red’) are allowed

osm2lanes: bool

If true, use experimental osm2lanes for figuring out lanes per road. If false, use the classic algorithm.

find_dog_legs_experiment: bool

Enable experimental dog-leg intersection merging

merge_osm_ways: Vec<OriginalRoad, Global>

Experimentally merge these OSM ways

Implementations

impl MapConfig

pub fn default_for_side(driving_side: DrivingSide) -> MapConfig