Struct map_model::Road

pub struct Road {
    pub id: RoadID,
    pub osm_tags: Tags,
    pub turn_restrictions: Vec<(RestrictionType, RoadID)>,
    pub complicated_turn_restrictions: Vec<(RoadID, RoadID)>,
    pub orig_id: OriginalRoad,
    pub speed_limit: Speed,
    pub access_restrictions: AccessRestrictions,
    pub zorder: isize,
    pub(crate) lanes_ltr: Vec<(LaneID, Direction, LaneType)>,
    pub center_pts: PolyLine,
    pub src_i: IntersectionID,
    pub dst_i: IntersectionID,
}

A Road represents a segment between exactly two Intersections. It contains Lanes as children.

Fields
id: RoadIDosm_tags: Tagsturn_restrictions: Vec<(RestrictionType, RoadID)>

self is 'from'

complicated_turn_restrictions: Vec<(RoadID, RoadID)>

self is 'from'. (via, to). Only BanTurns.

orig_id: OriginalRoadspeed_limit: Speedaccess_restrictions: AccessRestrictionszorder: isizelanes_ltr: Vec<(LaneID, Direction, LaneType)>

Invariant: A road must contain at least one child

center_pts: PolyLine

The physical center of the road, including sidewalks, after trimming. The order implies road orientation. No edits ever change this.

src_i: IntersectionIDdst_i: IntersectionID
Implementations

impl Road

pub fn lanes_ltr(&self) -> Vec<(LaneID, Direction, LaneType)>

Returns all lanes from the left side of the road to right. Left/right is determined by the orientation of center_pts.

pub fn get_left_side(&self, map: &Map) -> PolyLine

pub fn offset(&self, lane: LaneID) -> usize

Counting from the left side of the road

pub fn dir(&self, lane: LaneID) -> Direction

pub fn parking_to_driving(&self, parking: LaneID, map: &Map) -> Option<LaneID>

pub(crate) fn speed_limit_from_osm(&self) -> Speed

pub fn find_closest_lane<F: Fn(&Lane) -> bool>(
&self,
from: LaneID,
filter: F,
map: &Map
) -> Option<LaneID>

Includes off-side

pub fn all_lanes(&self) -> Vec<LaneID>

pub fn get_dir_change_pl(&self, map: &Map) -> PolyLine

This is the FIRST yellow line where the direction of the road changes. If multiple direction changes happen, the result is kind of arbitrary.

pub fn get_half_width(&self, map: &Map) -> Distance

pub fn get_thick_polygon(&self, map: &Map) -> Polygon

pub fn get_name(&self, lang: Option<&String>) -> String

pub fn get_rank(&self) -> RoadRank

pub fn all_bus_stops(&self, map: &Map) -> Vec<BusStopID>

pub fn percent_grade(&self, map: &Map) -> f64

Returns [-1.0, 1.0]. 0 is flat, positive is uphill, negative is downhill.

pub fn is_light_rail(&self) -> bool

pub fn is_footway(&self) -> bool

pub fn is_service(&self) -> bool

pub fn common_endpt(&self, other: &Road) -> IntersectionID

pub fn is_private(&self) -> bool

pub(crate) fn access_restrictions_from_osm(&self) -> AccessRestrictions

pub fn get_zone<'a>(&self, map: &'a Map) -> Option<&'a Zone>

impl Road

pub(crate) fn children_forwards(&self) -> Vec<(LaneID, LaneType)>

These are ordered from closest to center lane (left-most when driving on the right) to farthest (sidewalk)

pub(crate) fn children_backwards(&self) -> Vec<(LaneID, LaneType)>

pub(crate) fn dir_and_offset(&self, lane: LaneID) -> (Direction, usize)

lane must belong to this road. Offset 0 is the centermost lane on each side of a road, then it counts up from there.

pub(crate) fn children(&self, dir: Direction) -> Vec<(LaneID, LaneType)>

pub(crate) fn incoming_lanes(
&self,
i: IntersectionID
) -> Vec<(LaneID, LaneType)>

Returns lanes from the "center" going out

pub(crate) fn outgoing_lanes(
&self,
i: IntersectionID
) -> Vec<(LaneID, LaneType)>

Returns lanes from the "center" going out